postbro_backend/requirements.txt
Django>=5.0.0,<5.1.0
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
django-allauth>=0.61.1
django-environ>=0.11.2
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
django-filter>=23.5
Pillow>=10.2.0
python-dotenv>=1.0.1
stripe>=7.11.0
anthropic>=0.18.1
google-genai>=0.2.0
google-generativeai>=0.3.0
openai>=1.0.0
django-storages>=1.14.2
gunicorn>=21.2.0
whitenoise>=6.6.0
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
supabase>=2.0.0
dj-database-url>=2.1.0
yt-dlp>=2025.11.12
requests>=2.31.0
PyJWT>=2.8.0 
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
supabase>=2.0.0
dj-database-url>=2.1.0
yt-dlp>=2025.11.12
requests>=2.31.0
PyJWT>=2.8.0 